+++ app/gist/controller.js
@@ -2,32 +2,32 @@ import TwiddleResolver from "ember-twiddle/lib/twiddle-resolver";
 import File from "ember-twiddle/lib/file";
 
 export default Em.Controller.extend({
-  contentObserver: Em.observer('model.files.@each.compiled', function () {
-    Em.run.debounce(this, 'setupApplication', 500);
-  }.on('init')),
+  // contentObserver: Em.observer('model.files.@each.compiled', function () {
+  //   Em.run.debounce(this, 'setupApplication', 500);
+  // }.on('init')),
 
-  setupApplication () {
-    if(this.currentApp) {
-      Em.run(this.currentApp, 'destroy');
-    }
+  // setupApplication () {
+  //   if(this.currentApp) {
+  //     Em.run(this.currentApp, 'destroy');
+  //   }
 
-    var App = Em.Application.create({
-      name:         "App",
-      rootElement:  '#demo-app',
-      modulePrefix: 'demo-app',
-      Resolver:     TwiddleResolver.extend({files: this.get('model.files')})
-    });
+  //   var App = Em.Application.create({
+  //     name:         "App",
+  //     rootElement:  '#demo-app',
+  //     modulePrefix: 'demo-app',
+  //     Resolver:     TwiddleResolver.extend({files: this.get('model.files')})
+  //   });
 
-    this.currentApp = App;
-  },
+  //   this.currentApp = App;
+  // },
 
-  actions: {
-    addFile () {
-      this.get('model.files').pushObject(File.create());
-    },
+  // actions: {
+  //   addFile () {
+  //     this.get('model.files').pushObject(File.create());
+  //   },
 
-    removeFile (file) {
-      this.get('model.files').removeObject(file);
-    }
-  }
+  //   removeFile (file) {
+  //     this.get('model.files').removeObject(file);
+  //   }
+  // }
 });
