+++ app/gist/edit/template.hbs
@@ -1,0 +1,1 @@
+{{twiddle-editor model=model}}
