+++ app/gist/route.js
@@ -11,25 +11,6 @@ export default Em.Route.extend({
     });
   },
 
-  model () {
-    return this.get('github').createGist({
-      files: [
-        {name: 'application.hbs', content:
-          "Welcome to EmberTwiddle! {{foo}} {{outlet}}"
-        },
-        {name: 'index.hbs', content:
-          "Index is: {{bar}}"
-        },
-        {name: 'application/controller.js', content:
-          "export default Em.Controller.extend({foo: 'foo'});"
-        },
-        {name: 'controllers/index.js', content:
-          "export default Em.Controller.extend({bar: 'bar'});"
-        }
-      ]
-    });
-  },
-
   actions: {
     signInViaGithub () {
       this.get('session').open('github-oauth2').catch(function(error) {
