+++ app/router.js
@@ -7,6 +7,7 @@ var Router = Ember.Router.extend({
 
 Router.map(function() {
   this.route('gist', {path: '/'}, function() {
+    this.route('new', {path: '/'});
     this.route('edit', {path: '/:id'});
   });
 });
